mov test_json_field into test_fields_misc
