+++ odoo/addons/test_orm/tests/__init__.py
@@ -29,7 +29,6 @@ from . import (
     test_indexed_translation,
     test_inherits,
     test_ir_rules,
-    test_json_field,
     test_many2many,
     test_models,
     test_modules_db,
